backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# WebSocket support (included with FastAPI, but listed for clarity)
websockets==12.0

